requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-HTTPAuth==4.8.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pillow==10.2.0
Werkzeug==3.0.1
